requirements.txt
anthropic[vertex]
google-genai
openai
pyyaml
python-dotenv
requests
vllm==0.11.1